refactor(types): the types cast system changed
Now `CastTraits.hpp` contains only the base general `CastTraits` version and `CastTraits<>` specialization(which is contains the basic cast functions as `To()`, `UnsafeCast()`, `HierarchyCast()` and `IsInstanceOf()`).
All other specializations, like for `String` and numeric casts now located in their own separated header files.
`Cast.hpp` header now contains only `CastTraits.hpp` include PLUS `Cast<>` and `BasicCast` aliases
