+++ engine/types/cast/Cast.hpp
@@ -4,8 +4,6 @@
 namespace Ruby {
     using BasicCast = CastTraits<>;
 
-    using IntCast = CastTraits<Details::Cast::IntegralCastTag>;
-
     template<typename T, typename... Args>
     using Cast = CastTraits<T, Args...>;
 }
